Feature doctor endpoints

+++ src/main/java/com/ashish/hospital/hospitalManagement/service/impl/DoctorServiceImpl.java
@@ -1,12 +1,16 @@
 package com.ashish.hospital.hospitalManagement.service.impl;
 
-import com.ashish.hospital.hospitalManagement.dtos.doctor.DoctorCreateRequest;
-import com.ashish.hospital.hospitalManagement.dtos.doctor.DoctorResponse;
-import com.ashish.hospital.hospitalManagement.dtos.doctor.DoctorUpdateRequest;
+import com.ashish.hospital.hospitalManagement.dtos.doctor.*;
+import com.ashish.hospital.hospitalManagement.entity.Appointment;
+import com.ashish.hospital.hospitalManagement.entity.Department;
 import com.ashish.hospital.hospitalManagement.entity.Doctor;
 import com.ashish.hospital.hospitalManagement.entity.Patient;
 import com.ashish.hospital.hospitalManagement.exception.ResourceNotFoundException;
+import com.ashish.hospital.hospitalManagement.mapper.AppointmentMapper;
+import com.ashish.hospital.hospitalManagement.mapper.DepartmentMapper;
 import com.ashish.hospital.hospitalManagement.mapper.DoctorMapper;
+import com.ashish.hospital.hospitalManagement.repository.AppointmentRepository;
+import com.ashish.hospital.hospitalManagement.repository.DepartmentRepository;
 import com.ashish.hospital.hospitalManagement.repository.DoctorRepository;
 import com.ashish.hospital.hospitalManagement.service.DoctorService;
 import jakarta.persistence.EntityNotFoundException;
@@ -14,8 +18,8 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -24,13 +28,11 @@ public class DoctorServiceImpl implements DoctorService {
     private final DoctorRepository doctorRepository;
     private final DoctorMapper doctorMapper;
 
-    @Override
-    public DoctorResponse createDoctor(DoctorCreateRequest request) {
-        Doctor doctor = doctorMapper.toEntity(request);
-        Doctor savedDoctor = doctorRepository.save(doctor);
+    private final AppointmentRepository appointmentRepository;
+    private final AppointmentMapper appointmentMapper;
 
-        return doctorMapper.toResponse(savedDoctor);
-    }
+    private final DepartmentRepository departmentRepository;
+    private final DepartmentMapper departmentMapper;
 
     @Override
     public List<DoctorResponse> getAllDoctors() {
@@ -39,33 +41,58 @@ public class DoctorServiceImpl implements DoctorService {
     }
 
     @Override
-    public DoctorResponse getDoctorById(Long id) {
+    public DoctorDetailResponse getDoctorById(Long id) {
         Doctor doctor = doctorRepository.findById(id)
                 .orElseThrow(()-> new EntityNotFoundException("Doctor not found with id " + id));
-        return doctorMapper.toResponse(doctor);
+        return doctorMapper.toDetailResponse(doctor);
+    }
+
+    @Override
+    public List<DoctorAppointmentSummaryResponse> getAppointmentsByDoctorId(Long doctorId) {
+        List<Appointment> appointments = appointmentRepository.findByDoctorId(doctorId);
+
+        return appointmentMapper.toDoctorSummaryResponseList(appointments);
+    }
+
+    @Override
+    public List<DoctorDepartmentSummaryResponse> getDepartmentsByDoctorId(Long doctorId) {
+        List<Department> departments = departmentRepository.findByDoctorId(doctorId);
+
+        return departmentMapper.toDoctorSummaryResponseList(departments);
     }
 
     @Override
     @Transactional
-    public DoctorResponse updateDoctor(Long id, DoctorUpdateRequest updatedDoctor) {
+    public DoctorDetailResponse createDoctor(DoctorCreateRequest request) {
+        Doctor doctor = doctorMapper.toEntity(request);
+        Doctor savedDoctor = doctorRepository.save(doctor);
+
+        return doctorMapper.toDetailResponse(savedDoctor);
+    }
+
+    @Override
+    @Transactional
+    public DoctorDetailResponse updateDoctor(Long id, DoctorUpdateRequest request) {
         Doctor existing = doctorRepository.findById(id)
-                .orElseThrow(()-> new EntityNotFoundException("Doctor not found with id " + id));
-        doctorMapper.updateFromRequest(updatedDoctor, existing);
+                .orElseThrow(()-> new ResourceNotFoundException("Doctor not found with id " + id));
+        doctorMapper.updateFromRequest(request, existing);
 
         Doctor updated = doctorRepository.save(existing);
 
-        return doctorMapper.toResponse(updated);
+        return doctorMapper.toDetailResponse(updated);
     }
 
     @Override
+    @Transactional
     public void deleteDoctor(Long id) {
-        doctorRepository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Doctor not found with id " + id));
+        Doctor doctor = doctorRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Doctor not found with id " + id));
+
+        if(!doctor.getAppointments().isEmpty()){
+            throw new IllegalStateException("Doctor cannot be deleted because appointments exist. Cancel appointments first.");
+        }
+
         doctorRepository.deleteById(id);
     }
 
-    //TODO Once appointment impl is done
-    @Override
-    public List<Patient> getDoctorPatients(Long doctorId) {
-        return List.of();
-    }
 }
